+++ src/features/Calendar/index.ts
@@ -1,0 +1,1 @@
+export { default as CalendarTab } from './CalendarTab';
