+++ src/features/Games/index.ts
@@ -1,0 +1,1 @@
+export { default as GamesTab } from './GamesTab';
