+++ src/features/Players/PlayersTab.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import { Box, Button, Paper, TextField, Typography } from '@mui/material';
-import { loadPlayers, savePlayer } from '../lib/storage';
-import type { Player } from 'types/models';
-import theme from '../theme/theme';
+import { loadPlayers, savePlayer } from '@/lib/storage';
+import type { Player } from '@/types/models';
+import theme from '@/theme/theme';
 
-type PlayersProps = {
+type PlayersTabProps = {
     campaignId: string;
 };
 
-const Players: React.FC<PlayersProps> = ({ campaignId }) => {
+const PlayersTab: React.FC<PlayersTabProps> = ({ campaignId }) => {
     const [name, setName] = useState('');
     const [players, setPlayers] = useState<Player[]>([]);
 
@@ -60,7 +60,7 @@ const Players: React.FC<PlayersProps> = ({ campaignId }) => {
                     size="small"
                     sx={{
                         minWidth: '200px',
-                        width: '30vw',
+                        width: '30%',
                         maxWidth: '500px',
                     }}
                 />
@@ -79,6 +79,7 @@ const Players: React.FC<PlayersProps> = ({ campaignId }) => {
                     backgroundColor: 'rgba(0,0,0,0.1)',
                     minWidth: '150px',
                     overflowY: 'auto',
+                    height: 'min-content',
                 }}
             >
                 <Typography variant="h6" gutterBottom>
@@ -100,4 +101,4 @@ const Players: React.FC<PlayersProps> = ({ campaignId }) => {
     );
 };
 
-export default Players;
+export default PlayersTab;
