+++ src/features/Players/index.ts
@@ -1,0 +1,1 @@
+export { default as PlayersTab } from './PlayersTab';
