--- src/features/Stats.tsx
@@ -1,9 +1,0 @@
-const Stats = () => {
-    return (
-        <>
-            <h2>COMING SOON...</h2>
-        </>
-    );
-};
-
-export default Stats;
