+++ src/features/Trophy/index.ts
@@ -1,0 +1,1 @@
+export { default as TrophyTab } from './TrophyTab';
