+++ src/features/index.ts
@@ -1,0 +1,4 @@
+export { GamesTab } from './Games';
+export { PlayersTab } from './Players';
+export { TrophyTab } from './Trophy';
+export { CalendarTab } from './Calendar';
