+++ src/index.tsx
@@ -4,11 +4,14 @@ import './index.css';
 import App from './App.tsx';
 import { ThemeProvider } from '@emotion/react';
 import theme from './theme/theme.ts';
+import { CampaignProvider } from './context/CampaignContext.tsx';
 
 createRoot(document.getElementById('root')!).render(
     <StrictMode>
         <ThemeProvider theme={theme}>
-            <App />
+            <CampaignProvider>
+                <App />
+            </CampaignProvider>
         </ThemeProvider>
     </StrictMode>,
 );
